requirements.txt
# Autonomous Claude Agent - Python Dependencies
# ==============================================

# Web Framework
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
aiofiles==24.1.0
aiohttp==3.11.11

# Database & Vector Store
psycopg2-binary==2.9.11
qdrant-client==1.12.1
sentence-transformers==3.3.1
redis==5.2.1

# Configuration & Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
psutil==6.1.1
watchdog==6.0.0

# Browser Automation
playwright==1.49.1

# AWS SDK for SSM
boto3==1.35.0

